Decompile RIC func_8016DF74
